requirements.txt
numpy
pandas
matplotlib
datetime
yfinance
keras.models
streamlit
plotly
scikit-learn
tensorflow
